ml-service/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.2
tensorflow==2.14.0
PyMuPDF==1.23.8
opencv-python==4.8.1.78
python-multipart==0.0.6
python-dotenv==1.0.0
requests==2.31.0
aiofiles==23.2.1
Pillow==10.1.0
